Move logic of PullRequests and Pipeline to body
